requirements.txt
pytest==9.0.2
black==26.1.0
flake8==7.3.0
isort==7.0.0
python-dotenv==1.2.1
requests==2.32.5
jsonschema==4.26.0
allure-pytest==2.15.3